+++ msgflow/service/cli_service.py
@@ -1,6 +1,8 @@
 import sys
 from pydantic import BaseModel
-from typing import Any
+from typing import Any, Dict
+
+from ..bot import Bot
 
 
 class CliMessage:
@@ -13,11 +15,6 @@ class CliMessage:
     @property
     def text(self):
         return self._text
-
-    @property
-    def dialog_id(self) -> str:
-        # In CliService, a conversation is identified by the user's name
-        return self._user_name
 
     def respond(self, text):
         fmt = f"bot> {text}"
@@ -46,14 +43,14 @@ class CliService:
         self._out_fd = out_fd
 
     @classmethod
-    def from_config(cls, config: dict[str, object]):
+    def from_config(cls, config: Dict[str, object]):
         return cls(config=CliConfig(**config), in_fd=sys.stdin, out_fd=sys.stdout)
 
     def _show_prompt(self, user_name, out_fd):
         print(f"{user_name}> ", end="", file=out_fd)
         out_fd.flush()
 
-    def flow(self, bot):
+    def flow(self, bot: Bot):
         self._show_prompt(self._config.user_name, self._out_fd)
         for line in self._in_fd:
             text = line.rstrip("\n")
@@ -64,11 +61,11 @@ class CliService:
             msg = CliMessage(
                 text=text, user_name=self._config.user_name, out_fd=self._out_fd
             )
-            bot.handle(msg)
+            bot.handle(message=msg)
             self._show_prompt(self._config.user_name, self._out_fd)
 
         # print("", file=self._out_fd)
         print("Bye!", file=self._out_fd)
 
-    def post(self, text):
+    def post(self, text: str):
         print(text, file=self._out_fd)